LayoutTests/media/track/track-webvtt-tc021-valign-expected.txt

Tests cue vertical alignment (direction) from settings.


*** Testing text track 0
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].direction == 'vertical') OK
EXPECTED (cues[0].alignment == 'middle') OK
EXPECTED (cues[0].textPosition == '50') OK
EXPECTED (cues[1].direction == 'vertical-lr') OK
EXPECTED (cues[1].alignment == 'middle') OK
EXPECTED (cues[1].textPosition == '50') OK
EXPECTED (cues[2].direction == 'vertical') OK
EXPECTED (cues[2].alignment == 'start') OK
EXPECTED (cues[2].textPosition == '0') OK

*** Testing text track 1
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].direction == 'vertical') OK
EXPECTED (cues[0].alignment == 'middle') OK
EXPECTED (cues[0].textPosition == '50') OK
EXPECTED (cues[1].direction == 'vertical-lr') OK
EXPECTED (cues[1].alignment == 'middle') OK
EXPECTED (cues[1].textPosition == '50') OK
EXPECTED (cues[2].direction == 'vertical') OK
EXPECTED (cues[2].alignment == 'start') OK
EXPECTED (cues[2].textPosition == '0') OK

*** Testing text track 2
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].direction == 'horizontal') OK
EXPECTED (cues[1].direction == 'horizontal') OK
EXPECTED (cues[2].direction == 'horizontal') OK
END OF TEST
